+++ NEWS.md
@@ -1,5 +1,7 @@
 # colorpath 0.0.0 (development version)
 
+* Added rescaler functions: `rescaler_linear()` and `rescaler_bezier()`.
+
 * Added function `palette_bezier()` to create a palette function using a Bézier spline.
 
 * Added function `luv()` to convert HCL data frame to LUV matrix. (#2)